requirements.txt
functions-framework==3.*
flask>=3.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
google-genai>=1.0.0
google-auth>=2.0.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.16.0
google-cloud-documentai>=2.24.0
Pillow>=10.3.0
PyJWT>=2.8.0
requests>=2.31.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
twilio>=9.0.0
httpx>=0.27.0
python-multipart>=0.0.9
reportlab>=4.2.0
pdfplumber>=0.11.0
pymupdf>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
sentence-transformers>=3.0.0
# Email channel — Resend for outbound grade replies, imap_tools for the
# Zoho IMAP poller (cleaner API than stdlib imaplib for fetch + MOVE),
# pdf2image to render PDF attachment pages into JPEGs for the per-page
# grading flow.
resend>=2.0.0
imap-tools>=1.6.0
pdf2image>=1.17.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
